ajoute du helper get target User et refactor des fichiers et dossiers
