+++ apps/bot/src/discord/embed/branding.ts
