+++ apps/bot/src/discord/embed/create-op-embed.ts
