+++ apps/bot/src/discord/router.ts
@@ -4,7 +4,8 @@ import { devCommands } from '../domains/_dev/index.js';
 import { devilFruitCommands } from '../domains/devil_fruit/index.js';
 import { playerCommands } from '../domains/player/index.js';
 import { shipCommands } from '../domains/ship/commands/index.js';
-import type { Command } from '../shared/command.js';
+
+import type { Command } from './types.js';
 
 const allCommands = [...playerCommands, ...devilFruitCommands, ...devCommands, ...shipCommands];
 const registry = new Map<string, Command>(allCommands.map((c) => [c.name, c]));
