+++ apps/bot/src/domains/_dev/commands/embed.ts
@@ -1,5 +1,5 @@
-import type { Command } from '../../../shared/command.js';
-import { createOpEmbed } from '../../../shared/embed/create-op-embed.js';
+import { createOpEmbed } from '../../../discord/embed/create-op-embed.js';
+import type { Command } from '../../../discord/types.js';
 
 export const embedCommand: Command = {
   name: 'embed',
