+++ apps/bot/src/domains/player/commands/karma.ts
@@ -1,4 +1,4 @@
-import type { Command } from '../../../shared/command.js';
+import type { Command } from '../../../discord/types.js';
 import { getKarmaGrade } from '../karma.js';
 import { findOrCreatePlayer } from '../service.js';
