+++ apps/bot/src/shared/build-asset-url.ts
