--- apps/bot/src/shared/button-handler.ts
@@ -1,6 +1,0 @@
-import type { ButtonInteraction } from 'discord.js';
-
-export type ButtonHandler = {
-  customIdPrefix: string;
-  handle: (interaction: ButtonInteraction) => Promise<void>;
-};
