--- apps/bot/src/shared/command.ts
@@ -1,6 +1,0 @@
-import type { Message } from 'discord.js';
-
-export type Command = {
-  name: string;
-  handler: (message: Message, args: Array<string>) => Promise<void>;
-};
