+++ apps/bot/src/shared/utils.ts
